Capstone page complete
TO DO: (Quality of life updates)
=====
- Rework hamburger button to a slide in component that has the header links (Mobile only)
- Make a "return to top" button when the user scrolls down
- Create a "next project" and "previous project" buttons at the corners of each project page, allowing users to go to the next project whenever they want.
- Make a "Projects" button on the main header that takes you to the project list. (starting from capstone)
