+++ portfolio/package.json
@@ -17,6 +17,7 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@types/webpack-env": "^1.18.5",
     "autoprefixer": "^10.0.1",
     "daisyui": "^4.11.1",
     "eslint": "^8",
