+++ portfolio/src/app/components/HomePageHeader.tsx
@@ -2,12 +2,7 @@
 
 import Image from "next/image";
 import React from "react";
-import DropDown from "./DropDown";
 // import Link from "next/link";
-
-// HEADER MAY BE REWORKED INTO 2 DIFFERENT VERSIONS FOR THE WEBSITE
-// ONE WILL BE SITE WIDE AND THE OTHER WILL BE STRICTLY FOR THE HOME PAGE
-// THAT IS THE CURRENT PLAN
 
 export default function HeaderComponent() {
   return (
@@ -25,15 +20,7 @@ export default function HeaderComponent() {
           <p className="mt-3 text-sm lg:text-base">
             Hello! My name is Matthew Simms, a recent SAIT graduate from the
             Software Development program. I am primarily a Front-End React
-            Developer, based in {""}
-            <a
-              className="hover:underline"
-              href="https://www.google.ca/maps/place/Calgary,+AB/@51.0275069,-114.2529819,11z/data=!3m1!4b1!4m6!3m5!1s0x537170039f843fd5:0x266d3bb1b652b63a!8m2!3d51.0447331!4d-114.0718831!16zL20vMDFyMzI?entry=ttu"
-              target="_blank"
-            >
-              Calgary, AB
-            </a>
-            🍁🏒
+            Developer, based in Canada 🍁🏒
           </p>
         </div>
       </div>
