+++ portfolio/src/app/components/projects/Capstone.tsx
@@ -5,24 +5,21 @@ export default function TecTOC() {
   return (
     <div className="card xl:card-side bg-base-100 shadow-xl m-5 w-full hover:scale-110 transition duration-700">
       <figure className="xl:w-1/4">
-        <img
-          src="/images/JAWN-logo.png"
-          alt="TecTOC logo"
-        />
+        <img src="/images/JAWN-logo.png" alt="TecTOC logo" />
       </figure>
       <div className="card-body xl:w-3/4">
-        <h2 className="card-title">Schedule Distribution System (SDS)</h2>
+        <h2 className="card-title">SDS Capstone Prototype</h2>
         <p>
-        The <a className="italic">{'"Schedule Distribution System"'}</a> or
-          SDS for short, was our Capstone project. The idea was to give small
-          businesses as well as large corporations a more efficient method of
-          schedule distribution. Keeping all employees updated on the latest
-          changes.
+          The <a className="italic">{'"Schedule Distribution System"'}</a> or
+          SDS, was our Capstone project. The idea was to give small businesses
+          as well as large corporations a more efficient method of communicating
+          schedule updates, changes and allowing employees to make adjustments
+          through built in API.
         </p>
-        
+
         <div className="card-actions items-center justify-center md:justify-between">
-        <div className="flex-wrap flex flex-row justify-center">
-        <Image
+          <div className="flex-wrap flex flex-row justify-center">
+            <Image
               className="rounded-lg mr-3 my-2"
               src="/images/tailwindcss-logo.png"
               alt="TailwindCSS Logo"
@@ -61,28 +58,12 @@ export default function TecTOC() {
               width={50}
             />
           </div>
-          <div className="">
           <a
-            href="https://jawn-sds.vercel.app/api/auth/signin"
-            target="_blank"
-            className="btn btn-outline my-1 xl:mx-2 w-full xl:w-28"
-          >
-            Live Demo
-          </a>
-          <a
-            href="https://github.com/RCC-STUDIO/jawn-timetracker"
-            target="_blank"
-            className="btn btn-outline my-1 xl:mx-2 w-full xl:w-28 "
-          >
-            GitHub
-          </a>
-          {/*<a
             href="/pages/capstone"
-            className="btn btn-outline my-1 xl:mx-2 w-full xl:w-28"
+            className="btn btn-outline my-1 xl:mx-2 w-full xl:w-1/4"
           >
             Learn More
-          </a>*/}
-          </div>
+          </a>
         </div>
       </div>
     </div>
